better test for restoring receipts
Re-buying a non-consumable in a new install would present the user with
a native "would you like to redownload this for free?" dialog, but would
not count as a Restore, so the old code wouldn't fetch the original ID.

Now we just check if there's an old receipt fetch at all.

Tested by re-buying item on clean install, as well as restoring purchase
on clean install. Also tested that buying consumables work.